+++ packages/popover/src/components/PopoverBody.tsx
@@ -1,11 +1,11 @@
 /** @jsxImportSource @emotion/react */
+import * as Style from '../styles';
 import { createClassVariant } from '@jdesignlab/theme';
 import { POPOVER_BORDER_COLOR } from '../constants';
-import bodyStyle from '../styles/createBodyStyle';
 
 export const PopoverBody = (props: { children: React.ReactNode }) => {
   return (
-    <section className={createClassVariant('popover', 'content')} css={bodyStyle(POPOVER_BORDER_COLOR)}>
+    <section className={createClassVariant('popover', 'content')} css={Style.createBody(POPOVER_BORDER_COLOR)}>
       {props.children}
     </section>
   );
