+++ packages/popover/src/components/PopoverFooter.tsx
@@ -1,11 +1,10 @@
 /** @jsxImportSource @emotion/react */
+import * as Style from '../styles';
 import { createClassVariant } from '@jdesignlab/theme';
-import { POPOVER_BORDER_COLOR } from '../constants';
-import footerStyle from '../styles/createFooterStyle';
 
 export const popoverFooter = (props: { children: React.ReactNode }) => {
   return (
-    <footer className={createClassVariant('popover', 'footer')} css={footerStyle(POPOVER_BORDER_COLOR)}>
+    <footer className={createClassVariant('popover', 'footer')} css={Style.footer}>
       {props.children}
     </footer>
   );
