+++ packages/popover/src/components/PopoverHeader.tsx
@@ -1,11 +1,11 @@
 /** @jsxImportSource @emotion/react */
+import * as Style from '../styles';
 import { createClassVariant } from '@jdesignlab/theme';
 import { POPOVER_BORDER_COLOR } from '../constants';
-import headerStyle from '../styles/createHeaderStyle';
 
 export const PopoverHeader = (props: { children?: React.ReactNode }) => {
   return (
-    <header className={createClassVariant('popover', 'header')} css={headerStyle(POPOVER_BORDER_COLOR)}>
+    <header className={createClassVariant('popover', 'header')} css={Style.createHeader(POPOVER_BORDER_COLOR)}>
       {props.children}
     </header>
   );
