+++ packages/popover/src/components/PopoverTrigger.tsx
@@ -1,7 +1,7 @@
 /** @jsxImportSource @emotion/react */
+import * as Style from '../styles';
 import React, { useContext, Children, useEffect, useRef } from 'react';
 import { createClassVariant } from '@jdesignlab/theme';
-import triggerStyle from '../styles/createTriggerStyle';
 import usePopoverControl from '../hooks/usePopoverControl';
 import { PopoverContext } from '../context';
 
@@ -21,7 +21,7 @@ export const PopoverTrigger = (props: { children: React.ReactNode }) => {
     <div
       ref={triggerRef}
       className={createClassVariant('popover', 'trigger')}
-      css={triggerStyle}
+      css={Style.trigger}
       aria-pressed={context.isOpen}
       aria-expanded={context.isOpen}
       onClick={onTogglePopover}
