--- packages/popover/src/styles/createOverlayStyle.ts
@@ -1,11 +1,0 @@
-import { css } from '@emotion/react';
-
-const createOverlayStyle = css({
-  position: 'fixed',
-  top: 0,
-  left: 0,
-  width: '100vw',
-  height: '100vh'
-});
-
-export default createOverlayStyle;
