--- packages/popover/src/styles/createTriggerStyle.ts
@@ -1,10 +1,0 @@
-import { css } from '@emotion/react';
-
-const createTriggerStyle = () => {
-  return css({
-    display: 'inline-block',
-    position: 'relative'
-  });
-};
-
-export default createTriggerStyle;
